fix (comment): correct comment creation and DTO mapping refactor (services): move business logic from controllers to services

+++ node/express-prisma-typescript/src/domains/reaction/service/reaction.service.impl.ts
@@ -4,32 +4,37 @@ import { ReactionService } from './reaction.service'
 import { ReactionDto } from '../dto/reactionDto'
 import { isValidReactionType } from '@utils/validation'
 import { ConflictException, NotFoundException } from '@utils/errors'
-import { isEmpty } from 'class-validator'
+import { PostRepository } from '@domains/post/repository'
+import { UserRepository } from '@domains/user/repository'
+import { userExists } from '@utils/database'
 
 export class ReactionServiceImpl implements ReactionService {
-  constructor (private readonly repository: ReactionRepository) {}
+  constructor (private readonly repository: ReactionRepository, private readonly postRepository: PostRepository, private readonly userRepository: UserRepository) {}
 
   async addReaction (userId: string, postId: string, type: ReactionType): Promise<ReactionDto> {
-    if (isEmpty(postId)) { throw new NotFoundException('Post ID not provided') }
-
     if (!isValidReactionType(type)) {
       throw new ConflictException('Invalid reaction type')
     }
-
+    const postExists: boolean = await this.postRepository.existById(postId)
+    if (!postExists) { throw new NotFoundException('Post') }
     const reaction = await this.repository.addReaction(postId, userId, type)
 
     return new ReactionDto(reaction)
   }
 
   async removeReaction (userId: string, postId: string, type: ReactionType): Promise<void> {
+    if (!isValidReactionType(type)) { throw new ConflictException('Invalid reaction type') }
+
     await this.repository.removeReaction(postId, userId, type)
   }
 
   async getRetweetsByUser (userId: string): Promise<ReactionDto[]> {
+    if (!(await userExists(userId))) { throw new NotFoundException('User') }
     return await this.repository.findByUserAndType(userId, ReactionType.RETWEET)
   }
 
   async getLikesByUser (userId: string): Promise<ReactionDto[]> {
+    if (!(await userExists(userId))) { throw new NotFoundException('User') }
     return await this.repository.findByUserAndType(userId, ReactionType.LIKE)
   }
 }
